project/pyproject.toml
[project]
name = "carbon-aware-building-control"
version = "0.1.0"
description = "基于深度强化学习的建筑能源碳感知控制系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # EnergyPlus + Sinergym Environment
    "sinergym>=3.0.0",
    "eppy>=0.5.63",
    "geomeppy>=0.5.0",
    # RL Framework
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.0.0",
    # Data Processing
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.12.0",
    "pyyaml>=5.4.0",
    # Utilities
    "tqdm>=4.65.0",
    "tensorboard>=2.12.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    
    # Code quality
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]

